refactor(@xen-orchestra/backups): extract methods from Backup.js
getAdaptersByRemote , RemoteTimeoutError and runTasks
